requirements.txt
Flask
Flask-Sock
google-generativeai
python-dotenv
eventlet
